chore: 타입 위치 변경
